django/requirements.txt
amqp==5.0.5
asgiref==3.2.10
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
billiard==3.6.3.0
cachetools==4.1.1
celery==5.0.5
certifi==2019.9.11
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.1
django-storages==1.10.1
google-api-core==1.23.0
google-auth==1.22.1
google-cloud-core==1.4.3
google-cloud-storage==1.32.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
idna==2.10
img2pdf==0.4.0
importlib-metadata==3.4.0
kombu==5.0.2
lxml==4.5.2
multidict==5.1.0
pikepdf==1.19.3
Pillow==7.2.0
pipenv==2018.11.26
prompt-toolkit==3.0.16
protobuf==3.13.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyMySQL==1.0.2
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
soupsieve==2.2
sqlparse==0.3.1
typing-extensions==3.7.4.3
urllib3==1.25.11
vine==5.0.0
virtualenv==16.7.7
virtualenv-clone==0.5.3
wcwidth==0.2.5
whitenoise==5.2.0
yarl==1.6.3
zipp==3.4.0
